--- Login.py
@@ -1,111 +1,0 @@
-from PyQt5 import QtCore, QtGui, QtWidgets
-import source
-
-class Ui_Login(object):
-    def setupUi(self, Form):
-        Form.setObjectName("Form")
-        Form.resize(400, 560)
-        Form.setMinimumSize(QtCore.QSize(400, 560))
-        Form.setMaximumSize(QtCore.QSize(400, 560))
-        self.label = QtWidgets.QLabel(Form)
-        self.label.setGeometry(QtCore.QRect(60, 10, 281, 141))
-        self.label.setStyleSheet("image: url(:/newPrefix/logo mm.png);")
-        self.label.setText("")
-        self.label.setObjectName("label")
-        self.label_4 = QtWidgets.QLabel(Form)
-        self.label_4.setGeometry(QtCore.QRect(70, 160, 261, 20))
-        self.label_4.setLayoutDirection(QtCore.Qt.LeftToRight)
-        self.label_4.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_4.setObjectName("label_4")
-        self.lineEdit = QtWidgets.QLineEdit(Form)
-        self.lineEdit.setGeometry(QtCore.QRect(70, 260, 261, 31))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat Medium")
-        font.setPointSize(10)
-        self.lineEdit.setFont(font)
-        self.lineEdit.setObjectName("lineEdit")
-        self.label_3 = QtWidgets.QLabel(Form)
-        self.label_3.setGeometry(QtCore.QRect(70, 230, 261, 31))
-        self.label_3.setStyleSheet("font: 63 10pt \"Montserrat SemiBold\";")
-        self.label_3.setObjectName("label_3")
-        self.label_5 = QtWidgets.QLabel(Form)
-        self.label_5.setGeometry(QtCore.QRect(70, 290, 261, 31))
-        self.label_5.setStyleSheet("font: 63 10pt \"Montserrat SemiBold\";")
-        self.label_5.setObjectName("label_5")
-        self.lineEdit_2 = QtWidgets.QLineEdit(Form)
-        self.lineEdit_2.setGeometry(QtCore.QRect(70, 320, 261, 31))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat Medium")
-        font.setPointSize(10)
-        self.lineEdit_2.setFont(font)
-        self.lineEdit_2.setEchoMode(QtWidgets.QLineEdit.Password)
-        self.lineEdit_2.setObjectName("lineEdit_2")
-        self.label_6 = QtWidgets.QLabel(Form)
-        self.label_6.setGeometry(QtCore.QRect(70, 190, 261, 41))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat ExtraBold")
-        font.setPointSize(16)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_6.setFont(font)
-        self.label_6.setStyleSheet("color: rgb(57, 171, 121);")
-        self.label_6.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_6.setObjectName("label_6")
-        self.pushButton = QtWidgets.QPushButton(Form)
-        self.pushButton.setGeometry(QtCore.QRect(70, 380, 261, 31))
-        self.pushButton.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.pushButton.setMouseTracking(False)
-        self.pushButton.setStyleSheet("border-radius: 5px;\n"
-"font: 57 10pt \"Montserrat Medium\";\n"
-"background-color: rgb(57, 171, 121);\n"
-"color: rgb(255, 255, 255);")
-        self.pushButton.setObjectName("pushButton")
-        self.label_2 = QtWidgets.QLabel(Form)
-        self.label_2.setGeometry(QtCore.QRect(70, 530, 261, 20))
-        self.label_2.setLayoutDirection(QtCore.Qt.LeftToRight)
-        self.label_2.setAlignment(QtCore.Qt.AlignCenter)
-        self.label_2.setObjectName("label_2")
-        self.label_7 = QtWidgets.QLabel(Form)
-        self.label_7.setGeometry(QtCore.QRect(120, 430, 121, 20))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat Medium")
-        font.setPointSize(9)
-        self.label_7.setFont(font)
-        self.label_7.setObjectName("label_7")
-        self.label_8 = QtWidgets.QLabel(Form)
-        self.label_8.setGeometry(QtCore.QRect(240, 430, 51, 20))
-        font = QtGui.QFont()
-        font.setFamily("Montserrat Medium")
-        font.setPointSize(9)
-        font.setUnderline(True)
-        self.label_8.setFont(font)
-        self.label_8.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.label_8.setStyleSheet("color: rgb(0, 0, 255);")
-        self.label_8.setObjectName("label_8")
-
-        self.retranslateUi(Form)
-        QtCore.QMetaObject.connectSlotsByName(Form)
-
-    def retranslateUi(self, Form):
-        _translate = QtCore.QCoreApplication.translate
-        Form.setWindowTitle(_translate("Form", "Form"))
-        self.label_4.setText(_translate("Form", "By: @harismuhajir"))
-        self.lineEdit.setPlaceholderText(_translate("Form", "Masukkan username"))
-        self.label_3.setText(_translate("Form", "Username"))
-        self.label_5.setText(_translate("Form", "Password"))
-        self.lineEdit_2.setPlaceholderText(_translate("Form", "Masukkan password"))
-        self.label_6.setText(_translate("Form", "Masuk Akun"))
-        self.pushButton.setText(_translate("Form", "Masuk"))
-        self.label_2.setText(_translate("Form", "Money Manager © 2020 "))
-        self.label_7.setText(_translate("Form", "Belum punya akun? "))
-        self.label_8.setText(_translate("Form", "Daftar"))
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    Form = QtWidgets.QWidget()
-    ui = Ui_Login()
-    ui.setupUi(Form)
-    Form.show()
-    sys.exit(app.exec_())
